+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 -   Improve the documentation about noise simulations [#283](https://github.com/litebird/litebird_sim/pull/283)
 
+-   Use libinsdb to access the IMO [#282](https://github.com/litebird/litebird_sim/pull/282)
+
 -   Move from `flake8`/`black` to `ruff` [#281](https://github.com/litebird/litebird_sim/pull/281/)
 
 -   New module to simulate HWP systematics [#232](https://github.com/litebird/litebird_sim/pull/232)